Move MockObserver into observe crate and expose it optionally
Right now already and even more so in the future we need a couple of
observer types solely for testing purposes. We already have MockObserver
and DummyObserver, defined in different locations but with the former
pretty much subsuming all functionality provided by the latter.
This change moves MockObserver's definition into the observe crate and
exposes it to clients using it with the newly introduced 'test' feature
enabled. This step will allow us to have a single source for types
commonly used for testing.
