+++ rust/template/distributed_datalog/observe/Cargo.toml
@@ -2,3 +2,7 @@
 name = "observe"
 version = "0.1.0"
 edition = "2018"
+
+[features]
+# Expose additional types commonly used only for testing purposes.
+test = []
