+++ rust/template/distributed_datalog/observe/src/lib.rs
@@ -9,9 +9,14 @@
 
 mod observable;
 mod observer;
+#[cfg(any(test, feature = "test"))]
+mod test;
 
 pub use observable::Observable;
 pub use observable::UpdatesObservable;
 pub use observer::Observer;
 pub use observer::ObserverBox;
 pub use observer::SharedObserver;
+
+#[cfg(feature = "test")]
+pub use test::MockObserver;
