+++ rust/template/distributed_datalog/src/lib.rs
@@ -5,3 +5,6 @@ pub use observe::SharedObserver;
 pub use observe::UpdatesObservable;
 pub use tcp_channel::TcpReceiver;
 pub use tcp_channel::TcpSender;
+
+#[cfg(feature = "test")]
+pub use observe::MockObserver;
